requirements/common.txt
Django==2.2.4
psycopg2==2.7.3
psycopg2-binary==2.8.1
uwsgi==2.0.15
django_configurations==2.1
django_dotenv==1.4.1
djangorestframework==3.10.2
django-oauth-toolkit==1.2.0
django-filter==2.2.0
paho-mqtt==1.4.0
pid==2.2.5
protobuf==3.9.1
